refactor(FOROME-963): code review fixes
