+++ src/components/export-panel-modal.tsx
@@ -3,7 +3,7 @@ import { ReactElement } from 'react'
 import { ExportTypeEnum } from '@core/enum/export-type.enum'
 import { t } from '@i18n'
 import operationsStore from '@store/operations'
-import mainTableStore from '@store/ws/main-table'
+import mainTableStore from '@store/ws/main-table.store'
 import { showToast } from '@utils/notifications/showToast'
 import { MainTableDataCy } from './data-testid/main-table.cy'
 import { IPopperMenuProps, PopperMenu } from './popper-menu/popper-menu'
