+++ src/pages/filter/dtree/components/modals/components/modal-save-dataset.tsx
@@ -12,7 +12,7 @@ import dtreeStore from '@store/dtree'
 import filterStore from '@store/filter'
 import filterPresetsStore from '@store/filter-presets'
 import operations from '@store/operations'
-import mainTableStore from '@store/ws/main-table'
+import mainTableStore from '@store/ws/main-table.store'
 import zoneStore from '@store/ws/zone'
 import { Routes } from '@router/routes.enum'
 import { Button } from '@ui/button'
