+++ src/pages/filter/refiner/components/filter-refiner.tsx
@@ -1,7 +1,7 @@
 import { ReactElement, useEffect, useRef } from 'react'
 
 import useClientHeight from '@core/hooks/use-client-height'
-import mainTableStore from '@store/ws/main-table'
+import mainTableStore from '@store/ws/main-table.store'
 import { QueryBuilderGroups } from '../../common/groups/query-builder-groups'
 import { SelectedGroup } from './middle-column/selected-group'
 import { QuerySelected } from './right-column/query-selected'
