+++ src/pages/main/page.tsx
@@ -6,7 +6,7 @@ import datasetStore from '@store/dataset'
 import dirinfoStore from '@store/dirinfo'
 import dtreeStore from '@store/dtree'
 import filterStore from '@store/filter'
-import mainTableStore from '@store/ws/main-table'
+import mainTableStore from '@store/ws/main-table.store'
 import variantStore from '@store/ws/variant'
 import zoneStore from '@store/ws/zone'
 import { Header } from '@components/header'
