+++ src/pages/main/ui/info-list.tsx
@@ -7,7 +7,7 @@ import dirinfoStore from '@store/dirinfo'
 import { DocsList } from './docs-list'
 
 export const InfoList = observer((): ReactElement => {
-  const docs = get(datasetStore, 'dsinfo.doc', [])
+  const docs = get(datasetStore, 'dsInfo.doc', [])
   const baseDatasetName = dirinfoStore.ancestorsDsInfo[0][0]
 
   if (!docs[1]) return <Fragment />
