+++ src/pages/ws/ui/table/cell-gene.tsx
@@ -7,7 +7,7 @@ import { observer } from 'mobx-react-lite'
 import { ViewTypeEnum } from '@core/enum/view-type-enum'
 import { geneColorMap } from '@core/gene-color-map'
 import columnsStore from '@store/ws/columns'
-import mainTableStore from '@store/ws/main-table'
+import mainTableStore from '@store/ws/main-table.store'
 import { CellI } from './cell-interfaces'
 import { PlusIcon } from './plus-icon'
 import { RowHeight } from './table'
