+++ src/pages/ws/ui/table/table.tsx
@@ -18,7 +18,7 @@ import { t } from '@i18n'
 import datasetStore from '@store/dataset'
 import filterStore from '@store/filter'
 import columnsStore from '@store/ws/columns'
-import mainTableStore from '@store/ws/main-table'
+import mainTableStore from '@store/ws/main-table.store'
 import variantStore from '@store/ws/variant'
 import zoneStore from '@store/ws/zone'
 import { Routes } from '@router/routes.enum'
