+++ src/store/operations.ts
@@ -5,7 +5,7 @@ import { ExportTypeEnum } from '@core/enum/export-type.enum'
 import { getApiUrl } from '@core/get-api-url'
 import dtreeStore from '@store/dtree'
 import filterStore from '@store/filter'
-import mainTableStore from '@store/ws/main-table'
+import mainTableStore from '@store/ws/main-table.store'
 import zoneStore from '@store/ws/zone'
 import { Routes } from '@router/routes.enum'
 import {
